Fixed some compilation errors

+++ core/src/main/java/com/github/kill05/goldmountain/entity/HumanEntity.java
@@ -4,7 +4,7 @@ public interface HumanEntity extends Entity {
 
     int getId();
 
-    short getSpeed();
+    int getSpeed();
 
     PlayerCostume getCostume();
