+++ core/src/main/java/com/github/kill05/goldmountain/entity/Player.java
@@ -2,6 +2,11 @@ package com.github.kill05.goldmountain.entity;
 
 public interface Player extends HumanEntity {
 
-    int getTotalLevel();
+    /**
+     * Returns the player's level (unsigned int)
+     *
+     * @return the player's level
+     */
+    long getTotalLevel();
 
 }
